+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/SampleService.java
@@ -5,9 +5,9 @@ import java.util.Collection;
 import java.util.List;
 
 import uk.ac.bbsrc.tgac.miso.core.data.Barcodable.EntityType;
-import uk.ac.bbsrc.tgac.miso.core.service.SaveService;
 import uk.ac.bbsrc.tgac.miso.core.data.Sample;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleIdentity;
+import uk.ac.bbsrc.tgac.miso.core.service.SaveService;
 import uk.ac.bbsrc.tgac.miso.core.util.PaginatedDataSource;
 
 public interface SampleService
@@ -41,12 +41,6 @@ public interface SampleService
    * @throws IOException
    */
   Collection<Sample> listByIdList(List<Long> idList) throws IOException;
-
-  /**
-   * @return a List of all non-archived SampleTypes
-   * @throws IOException
-   */
-  Collection<String> listSampleTypes() throws IOException;
 
   Sample save(Sample sample) throws IOException;